Suggestion for clean build detection fix
The reason the old method doesn't work anymore is that the directory already exists because the entitlements file is put in there before this script is run. However, a project specific .h file is only put there afterwards which provides a way for us to check whether we're in a clean build.